+++ ghps/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 """ghps modules."""
 from .params import GHPS_VERSION
+from .errors import GHPSError, GHPSRuntimeError, GHPSValidationError
 from .server import GHPageServer
 __version__ = GHPS_VERSION
 
-__all__ = ["GHPageServer"]
+__all__ = ["GHPageServer", "GHPSError", "GHPSRuntimeError", "GHPSValidationError"]