+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `GHPSError` class
+- `GHPSValidationError` class
+- `GHPSRuntimeError` class
+### Changed
+- Test system modified
 ## [0.3] - 2026-04-05
 ### Added
 - `auto_open` parameter